Bug Fixed
When calculating the cascaded accrual, the summed accrual will be
assigned to corresponding category of balance. However, for the category
in balance, which is not found in accrual, will be assigned with
undefined.

When rendering the columns, the Col in persisted checks the type of
subs, and doesn't accept undefined, which causes error and crashes the
front end.

Solution:
For the category in balance, which is a leaf, and not found in accrual,
will be assigned with an empty list. Problem solved.